New new date picker layout
